+++ js/data.js
@@ -1,0 +1,41 @@
+const TITLE_VALUES = [
+  'The Ritz London',
+  'Rosewood London',
+  'The Langham, London',
+  'Baglioni Hotel London'
+];
+const TYPE_VALUES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+const TIME_VALUES = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+const FEATURE_VALUES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+const DESCRIPTION_VALUES = [
+  'Food and beverage credit to be used during your stay',
+  'Property credit to be used during your stay',
+  'Beverage to be used during your stay',
+  'Food to be used during your stay',
+  '$100 to be used during your stay'
+];
+const PHOTO_VALUES = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+
+export { TITLE_VALUES, TYPE_VALUES, TIME_VALUES, FEATURE_VALUES, DESCRIPTION_VALUES, PHOTO_VALUES };
